requirements.txt
Authlib==1.1.0
autopep8==2.0.0
bcrypt==4.0.1
bidict==0.22.0
blinker==1.5
cachelib==0.9.0
certifi==2022.9.24
cffi==1.15.1
charset-normalizer==2.1.1
click==8.1.3
cryptography==38.0.1
dnspython==2.2.1
email-validator==1.3.0
eventlet==0.33.1
Flask==2.2.2
Flask-Bcrypt==1.0.1
flask-dance==6.2.0
Flask-Login==0.6.2
Flask-Mail==0.9.1
Flask-Session==0.4.0
Flask-SocketIO==5.3.1
Flask-SQLAlchemy==3.0.2
Flask-WTF==1.0.1
gevent==22.10.2
GitHub-Flask==3.2.0
greenlet==2.0.1
h11==0.14.0
idna==3.4
importlib-metadata==5.0.0
itsdangerous==2.0.1
Jinja2==3.1.2
MarkupSafe==2.1.1
oauthlib==3.2.2
psycopg2-binary==2.9.4
pycodestyle==2.9.1
pycparser==2.21
python-dotenv==0.21.0
python-engineio==4.3.4
python-socketio==5.7.2
requests==2.28.1
requests-oauthlib==1.3.1
simple-websocket==0.8.1
six==1.16.0
SQLAlchemy==1.4.42
tinify==1.6.0
tomli==2.0.1
urllib3==1.26.12
URLObject==2.4.3
Werkzeug==2.2.2
wsproto==1.2.0
WTForms==3.0.1
zipp==3.9.0
zope.event==4.5.0
zope.interface==5.5.1
